+++ examples/multiFileProject/multiFileProject.h
@@ -1,0 +1,18 @@
+/****************************************************************************************************************************
+  multiFileProject.h
+  
+  For ESP32, ESP32_S2, ESP32_C3 boards with ESP32 core v2.0.0+
+  Written by Khoi Hoang
+
+  Built by Khoi Hoang https://github.com/khoih-prog/ESP32_PWM
+  Licensed under MIT license
+*****************************************************************************************************************************/
+
+// To demo how to include files in multi-file Projects
+
+#pragma once
+
+#define USING_MICROS_RESOLUTION       true    //false
+
+// Can be included as many times as necessary, without `Multiple Definitions` Linker Error
+#include "ESP32_PWM.hpp"